User ID Check in all controller

+++ ArtMarketPlaceAPI/Controllers/CartController.cs
@@ -16,6 +16,9 @@ namespace ArtMarketPlaceAPI.Controllers
         [HttpGet]
         public async Task<IActionResult> GetCustomerCart(int customerId)
         {
+            var currentUserId = User.FindFirst("id")?.Value;
+            if (currentUserId != customerId.ToString()) return Forbid();
+
             var cart = await _cartService.GetCartByUserIdAsync(customerId);
             return Ok(cart.MapToDto());
         }
@@ -25,6 +28,9 @@ namespace ArtMarketPlaceAPI.Controllers
         [HttpPost]
         public async Task<IActionResult> AddItemToCart(int customerId, int productId, int quantity)
         {
+            var currentUserId = User.FindFirst("id")?.Value;
+            if (currentUserId != customerId.ToString()) return Forbid();
+
             var cart = await _cartService.AddItemToCartAsync(customerId, productId, quantity);
             return Ok(cart.MapToDto());
         }
@@ -34,6 +40,11 @@ namespace ArtMarketPlaceAPI.Controllers
         [HttpPut("{id:int}")]
         public async Task<IActionResult> RemoveItemInCart(int cartId, int productId,int quantity)
         {
+            var cartToUpdate = await _cartService.GetCartByIdAsync(cartId);
+
+            var currentUserId = User.FindFirst("id")?.Value;
+            if (currentUserId != cartToUpdate.UserId.ToString()) return Forbid();
+
             var cart = await _cartService.RemoveItemFromCartAsync(cartId, productId, quantity);
             return Ok(cart.MapToDto());
         }
@@ -44,6 +55,11 @@ namespace ArtMarketPlaceAPI.Controllers
         [HttpDelete("{id:int}")]
         public async Task<IActionResult> DeleteCart(int cartId)
         {
+            var cart = await _cartService.GetCartByIdAsync(cartId);
+
+            var currentUserId = User.FindFirst("id")?.Value;
+            if (currentUserId != cart.UserId.ToString()) return Forbid();
+
             var isDeleted = await _cartService.DeleteCartAsync(cartId);
             if (!isDeleted) return NotFound("Cart not found!");
             return Ok("Cart deleted!");
